requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.0
alembic==1.13.1
celery==5.3.4
redis==5.0.1
stripe==7.8.0
sendgrid==6.11.0
reportlab==4.0.7
pandas==2.1.4
numpy==1.25.2
python-jose[cryptography]==3.3.0
httpx==0.25.2
Pillow==10.1.0
jinja2==3.1.2
aiofiles==23.2.1
python-slugify==8.0.1